Class DFreedom implemented.
This class is used to separate the solver from the particle.
This is much cleaner.
Since the solver acts identically on any degree of freedom,
it is not needed to repeat it again and again for each
degree of freedom.

The output is identical to the previous output, this means
that it works correctly.